requirement.txt
fastapi
uvicorn
sqlalchemy
passlib==1.7.4
bcrypt==4.0.1
python-jose
python-multipart